Allowed for decimals in playtime values

+++ src/main/java/me/aroze/uwucore/commands/PlaytimeCommand.java
@@ -24,10 +24,10 @@ public class PlaytimeCommand implements CommandExecutor {
 
 
         int ticks = (Bukkit.getOfflinePlayer(player)).getStatistic(Statistic.PLAY_ONE_MINUTE);
-        int seconds = ticks / 20;
-        int minutes = seconds / 60;
-        int hours = minutes / 60;
-        int days = hours / 24;
+        double seconds = ticks / 20;
+        double minutes = seconds / 60;
+        double hours = minutes / 60;
+        double days = hours / 24;
         sender.sendMessage(ChatUtils.color("\n&#947c9cPlaytime of &#cda2db" + player + "&#947c9c in different units:"));
         sender.sendMessage(ChatUtils.color("&#93c9c1 • " + ticks + " &#678580ticks"));
         sender.sendMessage(ChatUtils.color("&#93c9c1 • " + seconds + " &#678580seconds"));
